feat: a second pass of input.rs
